Annotate Jobs when observing them
Currently if a job gets into a Failed state and the is a fail history we will
keep sending JobError events to the server over and over.

This change adds the observed annotation to Jobs when we handle them to only
ever send a single even for a failed Job.

This is the same as done doe DaemonSets and Deployments so the code is
refactored to be more shareable and to avoid duplicating the annotation keys all
over.