Remove staging
As we have discontinued the concept of a staging environment the hardcoded
promotion flow of master -> dev -> staging -> prod no longer makes sense. A long
term change on how promotes are done is in the pipeline but until that is done
this change short circuits the promotion flow to skip staging.

This change also removes any mention of staging in the code base to avoid
confusion. Any code currently presenting something around the staging name is
also updated like the hamctl status command.