Show status of all environments in hamctl
Currently hamctl status only shows the status of hardcoded dev and prod
environments. This makes the command close to useless when used with multiple
environments of other names.

This change adds a new environments array to the server API which contains all
identified environments where a service is released to. These are then written
by hamctl to the user along with the hard coded ones.

For backwards compatibality the dev and prod fields are kept around to allow a
rolling upgrade of the server and hamctl versions to developers.