Make hamctl status output a list
This change updates the output of hamctl status to be more concise and remove
unneeded information. As the link to Jenkins and the vulnerability count is a
point in time datapoint it is likely to be outdated whenever something is
released.

This change makes the output a one line status of each environment.

    $ hamctl status
    Status for service svc

    dev:  2021-06-23 08:42:22 master-1234-5678 by Jane Doe: Useful bits
    prod: 2021-06-23 08:42:22 master-5678-1234 by Jane Doe: More useful bits