Replace home made mux with gorilla/mux (#407)
Currently the server runs a complicated set of mux operations to register
handlers and middlewares. Further more parsing of query parameters are done with
complicated  string parsing. All of this makes it hard to understand the API but
also hard to maintain.

This change migrates to the gorilla/mux module which makes path parsing
redundant and also allows for sub routed grouping of middlewares which makes for
a simpler registration flow.

Another motivation for the change is to allow the introduction of a prometheus
metrics middleware which works nicely with parameterized path segments. This
will be introduced at a later point.