Extract release notifiers into map (#409)
Currently we have a single NotifyReleaseHook function that calls all sorts of
things in one big mess. This makes it hard to reason about and also hard to
understand dependencies and boundaries.

This change extracts each of the parts of that function into a map creating a
clear boundary for each of the notification hooks. This simplifies the code and
also makes it simpler to extend in the future with eg. metrics.