Fix HTTP bug in ArtifactIDFromBranch lookup
Currently hamctl will use the ArtifactIDFromBranch action to look up what to
release when using the branch flag.

    hamctl release --service foo --branch bar --env baz

The action has a bug in the endpoint construction where it adds the branch name
twice. Once as a path segment and later as a query parameter. The server expects
only a query parameter. The bug was introduced way back in
1482e499a5f3fada4434e183fc0743145bb62019 (#195) but as the old HTTP server mux
was not that strict the bug wasn't discovered. With the change to using
gorilla/mux in bea4c3f975e07d6a129194137cedcdb3ce26f717 (#407) where paths needs
to match strictly things stopped working.