Inject copier instead of using globals
The current implementation of flows and git depends on a concrete global
implementation for copying in the copy package. This makes any usage hard to
test and also reason about. Further more does the copy functions rely on a
global logger instance making for an even more coupled implementation.

This change extracts the copying functionality into an injected struct with the
logger injected. No tests or functionality is changed. This is purely a
refactoring to enable future changes.