Remove global logger
This change removes the global logger which currently makes the codebase hard to
maintain. It requires magic log.Init() calls everywhere to ensure that the
logger is initialized correctly to avoid panics and makes dependencies unclear.

This change removes the global logger of the log package along with the
functions using it. Instead loggers are initialized top-level and injected
through out the code base.

Flag initialization of the logger might not work with this setup though as the
log config is used before flags are initialized.